Qt 6.6.3 source code

+++ qtwebengine/src/3rdparty/chromium/components/ntp_tiles/constants.h
@@ -1,4 +1,4 @@
-// Copyright 2016 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -7,15 +7,13 @@
 
 #include <stddef.h>
 
-#include <utility>
-
 namespace ntp_tiles {
 
 // Maximum number of custom links that can be set by the user. Used on desktop.
 extern const size_t kMaxNumCustomLinks;
 
 // Maximum number of Most Visited sites that will be generated. Used on desktop.
-extern const int kMaxNumMostVisited;
+extern const size_t kMaxNumMostVisited;
 
 // Maximum number of tiles that can be shown on the NTP.
 const int kMaxNumTiles = 10;
