+++ qtwebengine/src/3rdparty/chromium/ui/base/accelerators/global_media_keys_listener_win.cc
@@ -1,10 +1,10 @@
-// Copyright 2018 The Chromium Authors. All rights reserved.
+// Copyright 2018 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "ui/base/accelerators/global_media_keys_listener_win.h"
 
-#include "base/bind.h"
+#include "base/functional/bind.h"
 #include "base/metrics/histogram_macros.h"
 #include "ui/base/accelerators/accelerator.h"
 #include "ui/events/keycodes/keyboard_code_conversion_win.h"
@@ -66,7 +66,7 @@ void GlobalMediaKeysListenerWin::OnWndProc(HWND hwnd,
   // SingletonHwndHotKeyObservers should only send us hot key messages.
   DCHECK_EQ(WM_HOTKEY, static_cast<int>(message));
 
-  int win_key_code = HIWORD(lparam);
+  WORD win_key_code = HIWORD(lparam);
   KeyboardCode key_code = KeyboardCodeForWindowsKeyCode(win_key_code);
 
   // We should only receive hot key events for keys that we're observing.
